requirements.txt
flask
google-generativeai
PyPDF2
python-docx
python-pptx
python-dotenv
gTTS